feat(sandbox): simplify sandbox AST example
