+++ examples/sandbox/src/lib/functor.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react'
 import type * as T from '@traversable/registry'
-import { parseKey } from '@traversable/registry'
-import { fn } from '@traversable/registry'
+import { fn, parseKey } from '@traversable/registry'
 import { t } from '@traversable/schema'
 import { Json } from '@traversable/json'
 
@@ -17,6 +16,8 @@ import { set } from './set'
 import { map } from './map'
 import { spacemacs } from './theme'
 import { Hover } from './hover'
+
+const label = ['ᵃ', 'ᵇ', 'ᶜ', 'ᵈ', 'ᵉ', 'ᶠ', 'ᵍ', 'ʰ', 'ⁱ', 'ʲ', 'ᵏ', 'ˡ', 'ᵐ', 'ⁿ', 'ᵒ', 'ᵖ', '𐞥', 'ʳ', 'ˢ', 'ᵗ', 'ᵘ', 'ᵛ', 'ʷ', 'ˣ', 'ʸ', 'ᶻ'] as const
 
 export const Functor: T.Functor<Free, t.Schema> = {
   map(f) {
@@ -47,29 +48,6 @@ export const IndexedFunctor: T.Functor.Ix<t.Functor.Index, Free, Fixpoint> = {
 export const fold = fn.cata(Functor)
 export const unfold = fn.ana(Functor)
 export const foldWithIndex = fn.cataIx(IndexedFunctor)
-
-const $ts = {
-  typeName: (typeName: string) => <span style={{ color: spacemacs.var, fontStyle: 'italic' }}>{typeName}</span>,
-  typeKeyword: <span style={{ color: spacemacs.var, fontStyle: 'italic' }}>type </span>,
-  angle: {
-    lhs: <span style={{ color: spacemacs.head1 }}>{'<'}</span>,
-    rhs: <span style={{ color: spacemacs.head1 }}>{'>'}</span>,
-  },
-  brace: {
-    lhs: <span style={{ color: spacemacs.head1 }}>{'{'}</span>,
-    rhs: <span style={{ color: spacemacs.head1 }}>{'}'}</span>,
-  },
-  brack: {
-    lhs: <span style={{ color: spacemacs.head1 }}>{'['}</span>,
-    rhs: <span style={{ color: spacemacs.head1 }}>{']'}</span>,
-  },
-  paren: {
-    lhs: <span style={{ color: spacemacs.head1 }}>{'('}</span>,
-    rhs: <span style={{ color: spacemacs.head1 }}>{')'}</span>,
-  },
-  pipe: <span style={{ color: spacemacs.head1 }}>{' | '}</span>,
-  amp: <span style={{ color: spacemacs.head1 }}>{' & '}</span>,
-}
 
 const $text = (
   color: keyof typeof spacemacs,
@@ -139,82 +117,57 @@ const $js = {
   ttipBg: $text('ttip-bg'),
   ttipSl: $text('ttip-sl'),
   ttype: $text('ttype'),
-  type: ({ state = false, text }: { state?: boolean, text: string }) => () => <>{$text('type', { fontStyle: 'italic' })(text)}</>,
-  // type: $text('type'),
+  type: (text: string) => <>{$text('type', { fontStyle: 'italic' })(text)}</>,
   var: $text('var'),
   war: $text('war'),
   yellow: $text('yellow'),
   yellowBg: $text('yellow-bg'),
 }
 
-const base = (text: string) => <span style={{ color: spacemacs.base, fontWeight: 600 }}>{text}</span>
-
 export type TermWithTypeTree =
   | [React.ReactNode, React.ReactNode]
   | [TermWithTypeTree, type: TermWithTypeTree]
 
 export namespace Recursive {
-  export const toTypeString: T.Algebra<Free, string> = (x) => {
+  export const toHtml: T.IndexedAlgebra<t.Functor.Index, Free, TermWithTypeTree> = (x, path) => {
     switch (true) {
       default: return fn.exhaustive(x)
-      case x.tag === URI.set: return 'Set<' + x.def + '>'
-      case x.tag === URI.map: return 'Map<' + x.def + '>'
-      case t.isCore(x): return t.recurse.Recursive.toTypeString(x)
+      case x.tag === URI.never: return Html.Never(x)
+      case x.tag === URI.any: return Html.Any(x)
+      case x.tag === URI.unknown: return Html.Unknown(x)
+      case x.tag === URI.void: return Html.Void(x)
+      case x.tag === URI.null: return Html.Null(x)
+      case x.tag === URI.undefined: return Html.Undefined(x)
+      case x.tag === URI.symbol: return Html.Symbol(x)
+      case x.tag === URI.boolean: return Html.Boolean(x)
+      case x.tag === URI.integer: return Html.Integer(x)
+      case x.tag === URI.bigint: return Html.BigInt(x)
+      case x.tag === URI.number: return Html.Number(x)
+      case x.tag === URI.string: return Html.String(x)
+      case x.tag === URI.eq: return Html.Eq(x)
+      case x.tag === URI.array: return Html.Array(x, path)
+      case x.tag === URI.record: return Html.Record(x, path)
+      case x.tag === URI.optional: return Html.Optional(x, path)
+      case x.tag === URI.set: return Html.Set(x, path)
+      case x.tag === URI.map: return Html.Map(x, path)
+      case x.tag === URI.union: return Html.Union(x, path)
+      case x.tag === URI.intersect: return Html.Intersect(x, path)
+      case x.tag === URI.tuple: return Html.Tuple(x, path)
+      case x.tag === URI.object: return Html.Object(x, path)
     }
   }
 
-  export const toTypeHtml: T.Algebra<Free, React.ReactNode> = (x) => {
-    switch (true) {
-      default: return fn.exhaustive(x)
-      case x.tag === URI.never: return $js.type({ text: 'never' })() satisfies React.ReactNode
-      case x.tag === URI.any: return $js.type({ text: 'any' })() satisfies React.ReactNode
-      case x.tag === URI.unknown: return $js.type({ text: 'unknown' })() satisfies React.ReactNode
-      case x.tag === URI.void: return $js.type({ text: 'void' })() satisfies React.ReactNode
-      case x.tag === URI.null: return $js.type({ text: 'null' })() satisfies React.ReactNode
-      case x.tag === URI.undefined: return $js.type({ text: 'undefined' })() satisfies React.ReactNode
-      case x.tag === URI.symbol: return $js.type({ text: 'symbol' })() satisfies React.ReactNode
-      case x.tag === URI.boolean: return $js.type({ text: 'boolean' })() satisfies React.ReactNode
-      case x.tag === URI.integer: return $js.type({ text: 'integer' })() satisfies React.ReactNode
-      case x.tag === URI.bigint: return $js.type({ text: 'bigint' })() satisfies React.ReactNode
-      case x.tag === URI.number: return $js.type({ text: 'number' })() satisfies React.ReactNode
-      case x.tag === URI.string: return $js.type({ text: 'string' })() satisfies React.ReactNode
-      case x.tag === URI.array: return $line($js.const('('), x.def, $js.const(')'), $js.const('['), $js.const(']')) satisfies React.ReactNode
-      case x.tag === URI.record: return $line($js.h1('Record'), $js.aqua('<'), $js.type({ text: 'string' })(), $js.cursor(','), x.def, $js.aqua('>')) satisfies React.ReactNode
-      case x.tag === URI.optional: return $line(x.def, $js.blue(' | '), $js.type({ text: 'undefined' })()) satisfies React.ReactNode
-      case x.tag === URI.eq: return $line(x.def, $js.cblkLn('?')) satisfies React.ReactNode
-      case x.tag === URI.set: return $line($js.h1('Set'), $js.aqua('<'), x.def, $js.aqua('>')) satisfies React.ReactNode
-      case x.tag === URI.map: return $line($js.h1('Map'), $js.aqua('<'), x.def[0], base(', '), x.def[1], $js.aqua('>')) satisfies React.ReactNode
-      case x.tag === URI.union: return $line(x.def[0], ...x.def.slice(1).flatMap((y) => [$js.blue(' | '), y])) satisfies React.ReactNode
-      case x.tag === URI.intersect: return $line(x.def[0], ...x.def.slice(1).flatMap((y) => [$js.blue(' & '), y])) satisfies React.ReactNode
-      case x.tag === URI.tuple: return $line($ts.brack.lhs, x.def[0], ...x.def.slice(1).flatMap((y) => [$ts.typeName(', '), y]), $ts.brack.rhs) satisfies React.ReactNode
-      case x.tag === URI.object: {
-        const xs = Object.entries(x.def)
-        return xs.length === 0
-          ? $js.aqua('{}')
-          : $line(
-            $js.aqua('{ '),
-            ...xs.flatMap(([k, v], ix) => [
-              ...ix === 0 ? [] : [$js.cursor(', ')],
-              $js.prop(parseKey(k)),
-              $js.blue(x.opt.includes(k) ? '?: ' : ': '),
-              v,
-            ]),
-            $js.aqua(' }'),
-          ) satisfies React.ReactNode
-      }
-    }
-  }
-
-  const label = ['ᵃ', 'ᵇ', 'ᶜ', 'ᵈ', 'ᵉ', 'ᶠ', 'ᵍ', 'ʰ', 'ⁱ', 'ʲ', 'ᵏ', 'ˡ', 'ᵐ', 'ⁿ', 'ᵒ', 'ᵖ', '𐞥', 'ʳ', 'ˢ', 'ᵗ', 'ᵘ', 'ᵛ', 'ʷ', 'ˣ', 'ʸ', 'ᶻ'] as const
-
-  const jsonToHtml = Json.fold<React.ReactNode>((x) => {
+  export const jsonToHtml = Json.fold<React.ReactNode>((x) => {
     switch (true) {
       default: return fn.exhaustive(x)
       case x == null: return $js.magenta(`${x}`)
       case typeof x === 'string': return $line($js.base('"'), $js.aqua(x), $js.base('"'))
       case typeof x === 'number': return $js.meta(`${x}`)
       case typeof x === 'boolean': return $js.red(`${x}`)
-      case Json.isArray(x): return $line($js.aqua('['), ...x.flatMap((y, iy) => [...iy === 0 ? [] : [$js.cursor(', ')], y]), $js.aqua(']'))
+      case Json.isArray(x): return $line(
+        $js.aqua('['),
+        ...x.flatMap((y, iy) => [...iy === 0 ? [] : [$js.cursor(', ')], y]), $js.aqua(']'),
+      )
       case Json.isObject(x): {
         const xs = Object.entries(x)
         return xs.length === 0 ? $js.aqua('{}') : $line(
@@ -225,350 +178,432 @@ export namespace Recursive {
       }
     }
   })
+}
 
-  export const toTermWithTypeHtml: T.IndexedAlgebra<t.Functor.Index, Free, TermWithTypeTree> = (x, ix) => {
+namespace Html {
+  export function Never(_: t.never): TermWithTypeTree {
+    return [
+      $line($js.var('t'), $js.cursor('.'), $js.const('never')),
+      $js.type('never'),
+    ]
+  }
+
+  export function Any(_: t.any): TermWithTypeTree {
+    return [
+      $line($js.var('t'), $js.cursor('.'), $js.const('any')),
+      $js.type('any'),
+    ]
+  }
+
+  export function Unknown(_: t.unknown): TermWithTypeTree {
+    return [
+      $line($js.var('t'), $js.cursor('.'), $js.const('unknown')),
+      $js.type('unknown'),
+    ]
+  }
+
+  export function Void(_: t.void): TermWithTypeTree {
+    return [
+      $line($js.var('t'), $js.cursor('.'), $js.const('void')),
+      $js.type('void'),
+    ]
+  }
+
+  export function Null(_: t.null): TermWithTypeTree {
+    return [
+      $line($js.var('t'), $js.cursor('.'), $js.const('null')),
+      $js.type('null'),
+    ]
+  }
+
+  export function Undefined(_: t.undefined): TermWithTypeTree {
+    return [
+      $line($js.var('t'), $js.cursor('.'), $js.const('undefined')),
+      $js.type('undefined'),
+    ]
+  }
+
+  export function Symbol(_: t.symbol): TermWithTypeTree {
+    return [
+      $line($js.var('t'), $js.cursor('.'), $js.const('symbol')),
+      $js.type('symbol'),
+    ]
+  }
+
+  export function Boolean(_: t.boolean): TermWithTypeTree {
+    return [
+      $line($js.var('t'), $js.cursor('.'), $js.const('boolean')),
+      $js.type('boolean'),
+    ]
+  }
+
+  export function Integer(x: t.integer): TermWithTypeTree {
     switch (true) {
-      default: return fn.exhaustive(x)
-      case x.tag === URI.never: return [$line($js.var('t'), $js.cursor('.'), $js.const('never')), $js.type({ text: 'never' })()] satisfies TermWithTypeTree
-      case x.tag === URI.any: return [$line($js.var('t'), $js.cursor('.'), $js.const('any')), $js.type({ text: 'any' })()] satisfies TermWithTypeTree
-      case x.tag === URI.unknown: return [$line($js.var('t'), $js.cursor('.'), $js.const('unknown')), $js.type({ text: 'unknown' })()] satisfies TermWithTypeTree
-      case x.tag === URI.void: return [$line($js.var('t'), $js.cursor('.'), $js.const('void')), $js.type({ text: 'void' })()] satisfies TermWithTypeTree
-      case x.tag === URI.null: return [$line($js.var('t'), $js.cursor('.'), $js.const('null')), $js.type({ text: 'null' })()] satisfies TermWithTypeTree
-      case x.tag === URI.undefined: return [$line($js.var('t'), $js.cursor('.'), $js.const('undefined')), $js.type({ text: 'undefined' })()] satisfies TermWithTypeTree
-      case x.tag === URI.symbol: return [$line($js.var('t'), $js.cursor('.'), $js.const('symbol')), $js.type({ text: 'symbol' })()] satisfies TermWithTypeTree
-      case x.tag === URI.boolean: return [$line($js.var('t'), $js.cursor('.'), $js.const('boolean')), $js.type({ text: 'boolean' })()] satisfies TermWithTypeTree
-      case x.tag === URI.integer: switch (true) {
-        default: return [$line($js.var('t'), $js.cursor('.'), $js.const('integer')), $js.type({ text: 'number' })()] satisfies TermWithTypeTree
-        case isNumber(x.exclusiveMinimum):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('integer'), $js.cursor('.'), $js.const('moreThan'), $js.aqua('('), $js.const(`${x.exclusiveMinimum}`), $js.aqua(')')), $js.type({ text: 'number' })()] satisfies TermWithTypeTree
-        case isNumber(x.exclusiveMaximum):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('integer'), $js.cursor('.'), $js.const('lessThan'), $js.aqua('('), $js.const(`${x.exclusiveMaximum}`), $js.aqua(')')), $js.type({ text: 'number' })()] satisfies TermWithTypeTree
-        case isNumber(x.minimum):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('integer'), $js.cursor('.'), $js.const('min'), $js.aqua('('), $js.const(`${x.minimum}`), $js.aqua(')')), $js.type({ text: 'number' })()] satisfies TermWithTypeTree
-        case isNumber(x.maximum):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('integer'), $js.cursor('.'), $js.const('max'), $js.aqua('('), $js.const(`${x.maximum}`), $js.aqua(')')), $js.type({ text: 'number' })()] satisfies TermWithTypeTree
-      }
-      case x.tag === URI.bigint: switch (true) {
-        default: return [$line($js.var('t'), $js.cursor('.'), $js.const('bigint')), $js.type({ text: 'bigint' })()] satisfies TermWithTypeTree
-        case isBigInt(x.exclusiveMinimum):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('bigint'), $js.cursor('.'), $js.const('moreThan'), $js.aqua('('), $js.const(`${x.exclusiveMinimum}`), $js.aqua(')')), $js.type({ text: 'bigint' })()] satisfies TermWithTypeTree
-        case isBigInt(x.exclusiveMaximum):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('bigint'), $js.cursor('.'), $js.const('lessThan'), $js.aqua('('), $js.const(`${x.exclusiveMaximum}`), $js.aqua(')')), $js.type({ text: 'bigint' })()] satisfies TermWithTypeTree
-        case isBigInt(x.minimum):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('bigint'), $js.cursor('.'), $js.const('min'), $js.aqua('('), $js.const(`${x.minimum}`), $js.aqua(')')), $js.type({ text: 'bigint' })()] satisfies TermWithTypeTree
-        case isBigInt(x.maximum):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('bigint'), $js.cursor('.'), $js.const('max'), $js.aqua('('), $js.const(`${x.maximum}`), $js.aqua(')')), $js.type({ text: 'bigint' })()] satisfies TermWithTypeTree
-      }
-      case x.tag === URI.number: switch (true) {
-        default: return [$line($js.var('t'), $js.cursor('.'), $js.const('number')), $js.type({ text: 'number' })()] satisfies TermWithTypeTree
-        case isNumber(x.exclusiveMinimum):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('number'), $js.cursor('.'), $js.const('moreThan'), $js.aqua('('), $js.const(`${x.exclusiveMinimum}`), $js.aqua(')')), $js.type({ text: 'number' })()] satisfies TermWithTypeTree
-        case isNumber(x.exclusiveMaximum):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('number'), $js.cursor('.'), $js.const('lessThan'), $js.aqua('('), $js.const(`${x.exclusiveMaximum}`), $js.aqua(')')), $js.type({ text: 'number' })()] satisfies TermWithTypeTree
-        case isNumber(x.minimum):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('number'), $js.cursor('.'), $js.const('min'), $js.aqua('('), $js.const(`${x.minimum}`), $js.aqua(')')), $js.type({ text: 'number' })()] satisfies TermWithTypeTree
-        case isNumber(x.maximum):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('number'), $js.cursor('.'), $js.const('max'), $js.aqua('('), $js.const(`${x.maximum}`), $js.aqua(')')), $js.type({ text: 'number' })()] satisfies TermWithTypeTree
-      }
-      case x.tag === URI.string: switch (true) {
-        default: return [$line($js.var('t'), $js.cursor('.'), $js.const('string')), $js.type({ text: 'string' })()] satisfies TermWithTypeTree
-        case isNumber(x.minLength):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('string'), $js.cursor('.'), $js.const('min'), $js.aqua('('), $js.const(`${x.minLength}`), $js.aqua(')')), $js.type({ text: 'string' })()] satisfies TermWithTypeTree
-        case isNumber(x.maxLength):
-          return [$line($js.var('t'), $js.cursor('.'), $js.const('string'), $js.cursor('.'), $js.const('max'), $js.aqua('('), $js.const(`${x.maxLength}`), $js.aqua(')')), $js.type({ text: 'string' })()] satisfies TermWithTypeTree
-      }
-      case x.tag === URI.eq: {
-        const children = jsonToHtml(x.def as never)
-        return [
-          $line($js.var('t'), $js.cursor('.'), $js.const('eq'), $js.aqua('('), children, $js.aqua(')')),
-          children,
-        ] satisfies TermWithTypeTree
-      }
-      case x.tag === URI.array: return [
-        $line($js.var('t'), $js.cursor('.'), $js.const('array'), $js.aqua('('), <Hover texts={x.def} path={ix} />, $js.aqua(')')),
-        $line($js.const('('), x.def[1], $js.const(')'), $js.const('['), $js.const(']')),
-      ] satisfies TermWithTypeTree
-      case x.tag === URI.record: return [
-        $line($js.var('t'), $js.cursor('.'), $js.const('record'), $js.aqua('('), <Hover texts={x.def} path={ix} />, $js.aqua(')')),
-        $line($js.h1('Record'), $js.aqua('<'), $js.type({ text: 'string' })(), $js.cursor(', '), x.def[1], $js.aqua('>')),
-      ] satisfies TermWithTypeTree
-      case x.tag === URI.optional: return [
-        $line($js.var('t'), $js.cursor('.'), $js.const('optional'), $js.aqua('('), <Hover texts={x.def} path={ix} />, $js.aqua(')')),
-        $line(x.def[1], $js.blue(' | '), $js.type({ text: 'undefined' })()),
-      ] satisfies TermWithTypeTree
-      case x.tag === URI.set: return [
-        $line($js.var('t'), $js.cursor('.'), $js.const('set'), $js.aqua('('), <Hover texts={x.def} path={ix} />, $js.aqua(')')),
-        $line($js.h1('Set'), $js.aqua('<'), x.def[1], $js.aqua('>')),
-      ] satisfies TermWithTypeTree
-      case x.tag === URI.map: {
-        console.log(x.def)
+      default: return [
+        $line(
+          $js.var('t'),
+          $js.cursor('.'),
+          $js.const('integer'),
+        ),
+        $js.type('number'),
+      ]
+      case t.number(x.exclusiveMinimum): return [
+        $line(
+          $js.var('t'),
+          $js.cursor('.'),
+          $js.const('integer'),
+          $js.cursor('.'),
+          $js.const('moreThan'),
+          $js.aqua('('),
+          $js.const(`${x.exclusiveMinimum}`),
+          $js.aqua(')'),
+        ),
+        $js.type('number'),
+      ]
+      case t.number(x.exclusiveMaximum): return [
+        $line(
+          $js.var('t'),
+          $js.cursor('.'),
+          $js.const('integer'),
+          $js.cursor('.'),
+          $js.const('lessThan'),
+          $js.aqua('('),
+          $js.const(`${x.exclusiveMaximum}`),
+          $js.aqua(')'),
+        ),
+        $js.type('number'),
+      ]
+      case t.number(x.minimum): return [
+        $line(
+          $js.var('t'),
+          $js.cursor('.'),
+          $js.const('integer'),
+          $js.cursor('.'),
+          $js.const('min'),
+          $js.aqua('('),
+          $js.const(`${x.minimum}`),
+          $js.aqua(')'),
+        ),
+        $js.type('number'),
+      ]
+      case t.number(x.maximum): return [
+        $line(
+          $js.var('t'),
+          $js.cursor('.'),
+          $js.const('integer'),
+          $js.cursor('.'),
+          $js.const('max'),
+          $js.aqua('('),
+          $js.const(`${x.maximum}`),
+          $js.aqua(')'),
+        ),
+        $js.type('number'),
+      ]
+    }
+  }
+
+  export function BigInt(x: t.bigint): TermWithTypeTree {
+    switch (true) {
+      default: return [$line($js.var('t'), $js.cursor('.'), $js.const('bigint')), $js.type('bigint')]
+      case t.bigint(x.exclusiveMinimum): return [
+        $line(
+          $js.var('t'),
+          $js.cursor('.'),
+          $js.const('bigint'),
+          $js.cursor('.'),
+          $js.const('moreThan'),
+          $js.aqua('('),
+          $js.const(`${x.exclusiveMinimum}`),
+          $js.aqua(')'),
+        ),
+        $js.type('bigint'),
+      ]
+      case t.bigint(x.exclusiveMaximum): return [
+        $line(
+          $js.var('t'),
+          $js.cursor('.'),
+          $js.const('bigint'),
+          $js.cursor('.'),
+          $js.const('lessThan'),
+          $js.aqua('('),
+          $js.const(`${x.exclusiveMaximum}`),
+          $js.aqua(')'),
+        ),
+        $js.type('bigint'),
+      ]
+      case t.bigint(x.minimum): return [
+        $line(
+          $js.var('t'),
+          $js.cursor('.'),
+          $js.const('bigint'),
+          $js.cursor('.'),
+          $js.const('min'),
+          $js.aqua('('),
+          $js.const(`${x.minimum}`),
+          $js.aqua(')'),
+        ),
+        $js.type('bigint'),
+      ]
+      case t.bigint(x.maximum): return [
+        $line(
+          $js.var('t'),
+          $js.cursor('.'),
+          $js.const('bigint'),
+          $js.cursor('.'),
+          $js.const('max'),
+          $js.aqua('('),
+          $js.const(`${x.maximum}`),
+          $js.aqua(')'),
+        ),
+        $js.type('bigint'),
+      ]
+    }
+  }
+
+  export function Number(x: t.number): TermWithTypeTree {
+    switch (true) {
+      default: return [$line($js.var('t'), $js.cursor('.'), $js.const('number')), $js.type('number')]
+      case t.number(x.exclusiveMinimum):
         return [
           $line(
             $js.var('t'),
             $js.cursor('.'),
-            $js.const('map'),
+            $js.const('number'),
+            $js.cursor('.'),
+            $js.const('moreThan'),
             $js.aqua('('),
-            <Hover texts={x.def[0]} path={[...ix, 0]} />,
-            $js.cursor(', '),
-            <Hover texts={x.def[1]} path={[...ix, 1]} />,
-            $js.aqua(')')
-          ),
+            $js.const(`${x.exclusiveMinimum}`),
+            $js.aqua(')')),
+          $js.type('number'),
+        ]
+      case t.number(x.exclusiveMaximum):
+        return [
           $line(
-            $js.h1('Map'),
-            $js.aqua('<'),
-            x.def[0][1],
-            $js.base(', '),
-            x.def[1][1],
-            $js.aqua('>')
-          ),
-        ] satisfies TermWithTypeTree
-      }
-      case x.tag === URI.union: return [
-        $line(
-          $js.var('t'),
-          $js.cursor('.'),
-          $js.const('union'),
-          $js.aqua('('),
-          ...x.def.flatMap((y, iy) => [...iy === 0 ? [] : [$js.cursor(', ')], <Hover texts={y} path={[...ix, iy]} />]),
-          $js.aqua(')')
-        ),
-        $line(...x.def.flatMap((y, iy) => [...iy === 0 ? [] : [$js.blue(' | ')], y[1]])),
-      ] satisfies TermWithTypeTree
-      case x.tag === URI.intersect: return [
-        $line(
-          $js.var('t'),
-          $js.cursor('.'),
-          $js.const('intersect'),
-          $js.aqua('('),
-          ...x.def.flatMap((y, iy) => [...iy === 0 ? [] : [$js.cursor(', ')], <Hover texts={y} path={[...ix, iy]} />]),
-          $js.aqua(')')
-        ),
-        $line(...x.def.flatMap((y, iy) => [...iy === 0 ? [] : [$js.blue(' & ')], y[1]])),
-      ] satisfies TermWithTypeTree
-      case x.tag === URI.tuple: return [
-        $line(
-          $js.var('t'),
-          $js.cursor('.'),
-          $js.const('tuple'),
-          $js.aqua('('),
-          ...x.def.flatMap((y, iy) => [...iy === 0 ? [] : [$js.cursor(', ')], <Hover texts={y} path={[...ix, iy]} />]),
-          $js.aqua(')')
-        ),
-        $line(
-          $ts.brack.lhs,
-          ...x.def.flatMap((y, iy) => [
-            ...iy === 0 ? [] : [$js.cursor(', ')],
-            $js.const(label[iy] + (iy >= x.opt ? '?: ' : ': ')),
-            y[1]
-          ]),
-          $ts.brack.rhs,
-        ),
-      ] satisfies TermWithTypeTree
-      case x.tag === URI.object: {
-        const xs = Object.entries(x.def)
-        return xs.length === 0
-          ? [
-            $line($js.var('t'), $js.cursor('.'), $js.const('object'), $js.aqua('('), $js.brown('{}'), $js.aqua(')')),
-            $js.aqua('{}'),
-          ] satisfies TermWithTypeTree
-          : [
-            $line(
-              $js.var('t'),
-              $js.cursor('.'),
-              $js.const('object'),
-              $js.aqua('('),
-              $js.brown('{ '),
-              ...xs.flatMap(([k, v], iy) => [
-                ...iy === 0 ? [] : [$js.cursor(', ')],
-                $js.base(parseKey(k)),
-                $js.base(': '),
-                <Hover texts={v} path={[...ix, k]} />,
-              ]),
-              $js.brown(' }'),
-              $js.aqua(')'),
-            ),
-            $line(
-              $js.aqua('{ '),
-              ...xs.flatMap(([k, v], iy) => [
-                ...iy === 0 ? [] : [$js.cursor(', ')],
-                $js.prop(parseKey(k)),
-                $js.blue(x.opt.includes(k) ? '?: ' : ': '),
-                v[1],
-              ]),
-              $js.aqua(' }'),
-            ),
-          ] satisfies TermWithTypeTree
-      }
+            $js.var('t'),
+            $js.cursor('.'),
+            $js.const('number'),
+            $js.cursor('.'),
+            $js.const('lessThan'),
+            $js.aqua('('),
+            $js.const(`${x.exclusiveMaximum}`),
+            $js.aqua(')')),
+          $js.type('number'),
+        ]
+      case t.number(x.minimum):
+        return [
+          $line(
+            $js.var('t'),
+            $js.cursor('.'),
+            $js.const('number'),
+            $js.cursor('.'),
+            $js.const('min'),
+            $js.aqua('('),
+            $js.const(`${x.minimum}`),
+            $js.aqua(')')),
+          $js.type('number'),
+        ]
+      case t.number(x.maximum):
+        return [
+          $line(
+            $js.var('t'),
+            $js.cursor('.'),
+            $js.const('number'),
+            $js.cursor('.'),
+            $js.const('max'),
+            $js.aqua('('),
+            $js.const(`${x.maximum}`),
+            $js.aqua(')')),
+          $js.type('number'),
+        ]
     }
   }
 
-  export const toTermHtml: T.Algebra<Free, React.ReactNode> = (x) => {
+  export function String(x: t.string): TermWithTypeTree {
     switch (true) {
-      default: return fn.exhaustive(x)
-      case x.tag === URI.eq:
-      case x.tag === URI.never: return $line($js.var('t'), $js.cursor('.'), $js.const('never'))
-      case x.tag === URI.any: return $line($js.var('t'), $js.cursor('.'), $js.const('any'))
-      case x.tag === URI.unknown: return $line($js.var('t'), $js.cursor('.'), $js.const('unknown'))
-      case x.tag === URI.void: return $line($js.var('t'), $js.cursor('.'), $js.const('void'))
-      case x.tag === URI.null: return $line($js.var('t'), $js.cursor('.'), $js.const('null'))
-      case x.tag === URI.undefined: return $line($js.var('t'), $js.cursor('.'), $js.const('undefined'))
-      case x.tag === URI.symbol: return $line($js.var('t'), $js.cursor('.'), $js.const('symbol'))
-      case x.tag === URI.boolean: return $line($js.var('t'), $js.cursor('.'), $js.const('boolean'))
-      case x.tag === URI.integer: return $line($js.var('t'), $js.cursor('.'), $js.const('integer'))
-      case x.tag === URI.bigint: return $line($js.var('t'), $js.cursor('.'), $js.const('bigint'))
-      case x.tag === URI.number: return $line($js.var('t'), $js.cursor('.'), $js.const('number'))
-      case x.tag === URI.string: return $line($js.var('t'), $js.cursor('.'), $js.const('string'))
-      case x.tag === URI.array: return $line($js.var('t'), $js.cursor('.'), $js.const('array'), $js.aqua('('), x.def, $js.aqua(')'))
-      case x.tag === URI.record: return $line($js.var('t'), $js.cursor('.'), $js.const('record'), $js.aqua('('), x.def, $js.aqua(')'))
-      case x.tag === URI.optional: return $line($js.var('t'), $js.cursor('.'), $js.const('optional'), $js.aqua('('), x.def, $js.aqua(')'))
-      case x.tag === URI.set: return $line($js.var('t'), $js.cursor('.'), $js.const('set'), $js.aqua('('), x.def, $js.aqua(')'))
-      case x.tag === URI.map: return $line($js.var('t'), $js.cursor('.'), $js.const('map'), $js.aqua('('), x.def[0], $js.cursor(', '), x.def[1], $js.aqua(')'))
-      case x.tag === URI.union: return $line(
+      default: return [$line($js.var('t'), $js.cursor('.'), $js.const('string')), $js.type('string')]
+      case t.number(x.minLength):
+        return [
+          $line(
+            $js.var('t'),
+            $js.cursor('.'),
+            $js.const('string'),
+            $js.cursor('.'),
+            $js.const('min'),
+            $js.aqua('('),
+            $js.const(`${x.minLength}`),
+            $js.aqua(')')),
+          $js.type('string'),
+        ]
+      case t.number(x.maxLength):
+        return [
+          $line(
+            $js.var('t'),
+            $js.cursor('.'),
+            $js.const('string'),
+            $js.cursor('.'),
+            $js.const('max'),
+            $js.aqua('('),
+            $js.const(`${x.maxLength}`),
+            $js.aqua(')')),
+          $js.type('string'),
+        ]
+    }
+  }
+
+  export function Eq<T>(x: t.eq<T>): TermWithTypeTree {
+    const children = Recursive.jsonToHtml(x.def as never)
+    return [
+      $line($js.var('t'), $js.cursor('.'), $js.const('eq'), $js.aqua('('), children, $js.aqua(')')),
+      children,
+    ]
+  }
+
+  export function Array<T extends TermWithTypeTree>(x: t.array<T>, path: (keyof any)[]): TermWithTypeTree {
+    return [
+      $line($js.var('t'), $js.cursor('.'), $js.const('array'), $js.aqua('('), <Hover texts={x.def} path={path} />, $js.aqua(')')),
+      $line($js.const('('), x.def[1], $js.const(')'), $js.const('['), $js.const(']')),
+    ]
+  }
+
+  export function Record<T extends TermWithTypeTree>(x: t.record<T>, path: (keyof any)[]): TermWithTypeTree {
+    return [
+      $line($js.var('t'), $js.cursor('.'), $js.const('record'), $js.aqua('('), <Hover texts={x.def} path={path} />, $js.aqua(')')),
+      $line($js.h1('Record'), $js.aqua('<'), $js.type('string'), $js.cursor(', '), x.def[1], $js.aqua('>')),
+    ]
+  }
+
+  export function Optional<T extends TermWithTypeTree>(x: t.optional<T>, path: (keyof any)[]): TermWithTypeTree {
+    return [
+      $line($js.var('t'), $js.cursor('.'), $js.const('optional'), $js.aqua('('), <Hover texts={x.def} path={path} />, $js.aqua(')')),
+      $line(x.def[1], $js.blue(' | '), $js.type('undefined')),
+    ]
+  }
+
+  export function Set<T extends TermWithTypeTree>(x: set<T>, path: (keyof any)[]): TermWithTypeTree {
+    return [
+      $line($js.var('t'), $js.cursor('.'), $js.const('set'), $js.aqua('('), <Hover texts={x.def} path={path} />, $js.aqua(')')),
+      $line($js.h1('Set'), $js.aqua('<'), x.def[1], $js.aqua('>')),
+    ]
+  }
+
+  export function Map<K extends TermWithTypeTree, V extends TermWithTypeTree>(x: map<K, V>, path: (keyof any)[]): TermWithTypeTree {
+    return [
+      $line(
+        $js.var('t'),
+        $js.cursor('.'),
+        $js.const('map'),
+        $js.aqua('('),
+        <Hover texts={x.def[0]} path={[...path, 0]} />,
+        $js.cursor(', '),
+        <Hover texts={x.def[1]} path={[...path, 1]} />,
+        $js.aqua(')')
+      ),
+      $line($js.h1('Map'), $js.aqua('<'), x.def[0][1], $js.base(', '), x.def[1][1], $js.aqua('>')),
+    ]
+  }
+
+  export function Union<T extends readonly TermWithTypeTree[]>(xs: t.union<T>, path: (keyof any)[]): TermWithTypeTree {
+    return [
+      $line(
         $js.var('t'),
         $js.cursor('.'),
         $js.const('union'),
         $js.aqua('('),
-        ...x.def.flatMap((y, iy) => [...iy === 0 ? [] : [$js.cursor(', ')], y]),
-        $js.aqua(')'),
-      )
-      case x.tag === URI.intersect: return $line(
+        ...xs.def.flatMap((x, ix) => [
+          ...ix === 0 ? [] : [$js.cursor(', ')],
+          <Hover texts={x} path={[...path, ix]} />
+        ]),
+        $js.aqua(')')
+      ),
+      $line(...xs.def.flatMap((x, ix) => [...ix === 0 ? [] : [$js.blue(' | ')], x[1]])),
+    ]
+  }
+
+  export function Intersect<T extends readonly TermWithTypeTree[]>(xs: t.intersect<T>, path: (keyof any)[]): TermWithTypeTree {
+    return [
+      $line(
         $js.var('t'),
         $js.cursor('.'),
         $js.const('intersect'),
         $js.aqua('('),
-        ...x.def.flatMap((y, iy) => [...iy === 0 ? [] : [$js.cursor(', ')], y]),
-        $js.aqua(')'),
-      )
-      case x.tag === URI.tuple: return $line(
+        ...xs.def.flatMap((x, ix) => [
+          ...ix === 0 ? [] : [$js.cursor(', ')],
+          <Hover texts={x} path={[...path, ix]} />
+        ]),
+        $js.aqua(')')
+      ),
+      $line(...xs.def.flatMap((x, ix) => [...ix === 0 ? [] : [$js.blue(' & ')], x[1]])),
+    ]
+  }
+
+  export function Tuple<T extends readonly TermWithTypeTree[]>(xs: t.tuple<T>, path: (keyof any)[]): TermWithTypeTree {
+    return [
+      $line(
         $js.var('t'),
         $js.cursor('.'),
         $js.const('tuple'),
         $js.aqua('('),
-        ...x.def.flatMap((y, iy) => [...iy === 0 ? [] : [$js.cursor(', '), y]]),
-        $js.aqua(')'),
-      )
-      case x.tag === URI.object: {
-        const xs = Object.entries(x.def)
-        return xs.length === 0
-          ? $line(
-            $js.var('t'),
-            $js.cursor('.'),
-            $js.const('object'),
-            $js.aqua('('),
-            $js.brown('{}'),
-            $js.aqua(')')
-          )
-          : $line(
-            $js.var('t'),
-            $js.cursor('.'),
-            $js.const('object'),
-            $js.aqua('('),
-            $js.brown('{ '),
-            ...xs.flatMap(([k, v], iy) => [
-              ...iy === 0 ? [] : [$js.cursor(', ')],
-              $js.base(parseKey(k)),
-              $js.base(': '),
-              v,
-            ]),
-            $js.brown(' }'),
-            $js.aqua(')'),
-          )
-      }
-    }
+        ...xs.def.flatMap((x, ix) => [
+          ...ix === 0 ? [] : [$js.cursor(', ')],
+          <Hover texts={x} path={[...path, ix]} />
+        ]),
+        $js.aqua(')')
+      ),
+      $line(
+        $js.h1('['),
+        ...xs.def.flatMap((x, ix) => [
+          ...ix === 0 ? [] : [$js.cursor(', ')],
+          $js.const(label[ix] + (ix >= xs.opt ? '?: ' : ': ')),
+          x[1]
+        ]),
+        $js.h1(']'),
+      ),
+    ]
   }
 
-  const jsonToString = Json.fold<string>((x) => {
-    switch (true) {
-      default: return fn.exhaustive(x)
-      case typeof x === 'string': return `"${x}"`
-      case Json.isScalar(x): return `${x}`
-      case Json.isArray(x): return '[' + x.join(', ') + ']'
-      case Json.isObject(x): {
-        const xs = Object.entries(x)
-        return xs.length === 0 ? '{}' : '{ ' + xs.map(([k, v]) => `${parseKey(k)}: ${v}`).join(', ') + ' }'
-      }
-    }
-  })
-
-  const isNumber = (u: unknown): u is number => typeof u === 'number'
-  const isBigInt = (u: unknown): u is bigint => typeof u === 'bigint'
-
-  export const toSchemaString: T.Algebra<Free, string> = (x) => {
-    switch (true) {
-      default: return fn.exhaustive(x)
-      case x.tag === URI.any: return 't.any'
-      case x.tag === URI.never: return 't.never'
-      case x.tag === URI.unknown: return 't.unknown'
-      case x.tag === URI.void: return 't.void'
-      case x.tag === URI.null: return 't.null'
-      case x.tag === URI.undefined: return 't.undefined'
-      case x.tag === URI.symbol: return 't.symbol'
-      case x.tag === URI.boolean: return 't.boolean'
-      case x.tag === URI.integer: switch (true) {
-        case isNumber(x.exclusiveMinimum): return 't.integer.moreThan(' + x.exclusiveMinimum + ')'
-        case isNumber(x.exclusiveMaximum): return 't.integer.lessThan(' + x.exclusiveMaximum + ')'
-        case isNumber(x.minimum): return 't.integer.min(' + x.minimum + ')'
-        case isNumber(x.maximum): return 't.integer.max(' + x.maximum + ')'
-        default: return 't.integer'
-      }
-      case x.tag === URI.bigint: switch (true) {
-        case isBigInt(x.exclusiveMinimum): return 't.bigint.moreThan(' + x.exclusiveMinimum + ')'
-        case isBigInt(x.exclusiveMaximum): return 't.bigint.lessThan(' + x.exclusiveMinimum + ')'
-        case isBigInt(x.minimum): return 't.bigint.min(' + x.minimum + ')'
-        case isBigInt(x.maximum): return 't.bigint.max(' + x.maximum + ')'
-        default: return 't.bigint'
-      }
-      case x.tag === URI.number: switch (true) {
-        case isNumber(x.exclusiveMinimum): return 't.number.moreThan(' + x.exclusiveMinimum + ')'
-        case isNumber(x.exclusiveMaximum): return 't.number.lessThan(' + x.exclusiveMaximum + ')'
-        case isNumber(x.minimum): return 't.number.min(' + x.minimum + ')'
-        case isNumber(x.maximum): return 't.number.max(' + x.maximum + ')'
-        default: return 't.number'
-      }
-      case x.tag === URI.string: switch (true) {
-        case isNumber(x.minLength): return 't.string.min(' + x.minLength + ')'
-        case isNumber(x.maxLength): return 't.string.max(' + x.maxLength + ')'
-        default: return 't.string'
-      }
-      case x.tag === URI.string: return 't.string'   // TODO
-      case x.tag === URI.eq: return 't.eq(' + jsonToString(x.def) + ')'
-      case x.tag === URI.optional: return 't.optional(' + x.def + ')'
-      case x.tag === URI.array: return 't.array(' + x.def + ')'
-      case x.tag === URI.record: return 't.record(' + x.def + ')'
-      case x.tag === URI.enum as never: return 't.enum(' + Object.values(x.def).join(', ') + ')'
-      case x.tag === URI.nonnullable as never: return 't.enum(' + Object.values(x.def).join(', ') + ')'
-      case x.tag === URI.union: return 't.union(' + x.def.join(', ') + ')'
-      case x.tag === URI.intersect: return 't.intersect(' + x.def.join(', ') + ')'
-      case x.tag === URI.tuple: return 't.tuple(' + x.def.join(', ') + ')'
-      case x.tag === URI.set: return 't.set(' + x.def + ')'
-      case x.tag === URI.map: return 't.map(' + x.def[0] + ', ' + x.def[1] + ')'
-      case x.tag === URI.object: {
-        const xs = Object.entries(x.def)
-        return xs.length === 0 ? 't.object({})' : 't.object({ ' + xs.map(([k, v]) => `${parseKey(k)}: ${v}`).join(', ') + ' })'
-      }
-    }
+  export function Object<T extends { [x: string]: TermWithTypeTree }>(x: t.object<T>, path: (keyof any)[]): TermWithTypeTree {
+    const xs = globalThis.Object.entries(x.def)
+    const opt: string[] = x.opt as never
+    return xs.length === 0 ? [
+      $line(
+        $js.var('t'),
+        $js.cursor('.'),
+        $js.const('object'),
+        $js.aqua('('),
+        $js.brown('{}'),
+        $js.aqua(')')
+      ),
+      $js.aqua('{}'),
+    ] : [
+      $line(
+        $js.var('t'),
+        $js.cursor('.'),
+        $js.const('object'),
+        $js.aqua('('),
+        $js.brown('{ '),
+        ...xs.flatMap(([k, v], iy) => [
+          ...iy === 0 ? [] : [$js.cursor(', ')],
+          $js.base(parseKey(k)),
+          $js.base(': '),
+          <Hover texts={v} path={[...path, k]} />,
+        ]),
+        $js.brown(' }'),
+        $js.aqua(')'),
+      ),
+      $line(
+        $js.aqua('{ '),
+        ...xs.flatMap(([k, v], iy) => [
+          ...iy === 0 ? [] : [$js.cursor(', ')],
+          $js.prop(parseKey(k)),
+          $js.blue(opt.includes(k) ? '?: ' : ': '),
+          v[1],
+        ]),
+        $js.aqua(' }'),
+      ),
+    ]
   }
 }
 
-export const toTypeString = fold(Recursive.toTypeString)
-export const toSchemaString = fold(Recursive.toSchemaString)
-
-export const toTypeHtml
-  : <S extends t.Schema>(term: S) => React.JSX.Element
-  = (term) => {
-    const ui = fold(Recursive.toTypeHtml)(term)
-    const Elem = isReact.FC(ui) ? ui : () => ui
-    return <pre style={{ position: 'relative' }}><Elem /></pre>
-  }
-
-export const toTermHtml
-  : <S extends t.Schema>(term: S) => React.JSX.Element
-  = (term) => {
-    return <pre>
-      {fold(Recursive.toTermHtml)(term)}
-    </pre>
-  }
-
-export const toTermWithTypeHtml
+export const toHtml
   : <S extends t.Schema>(term: S, ix?: t.Functor.Index) => TermWithTypeTree
-  = (term, ix) => foldWithIndex(Recursive.toTermWithTypeHtml)(term as never, ix ?? [])
+  = (term, ix) => foldWithIndex(Recursive.toHtml)(term as never, ix ?? [])
